+++ Commands/Provisioning/Tenant/GetTenantSequenceSite.cs
@@ -1,4 +1,5 @@
-﻿using OfficeDevPnP.Core.Framework.Provisioning.Model;
+﻿#if !ONPREMISES
+using OfficeDevPnP.Core.Framework.Provisioning.Model;
 using SharePointPnP.PowerShell.CmdletHelpAttributes;
 using SharePointPnP.PowerShell.Commands.Base.PipeBinds;
 using System;
@@ -43,3 +44,4 @@ namespace SharePointPnP.PowerShell.Commands.Provisioning.Tenant
         }
     }
 }
+#endif